test: [FE] ControlButton에 대한 테스트 코드 변경 ControlButton에 대한 테스트 코드 변경
